weather_bot/requirements.txt
requests
python-dotenv
pyTelegramBotAPI
flask
gunicorn
